+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './currency';
 export * from './toast';
 export * from './error';
 export * from './string';
+export * from './route-metadata';